moved the mock class to the test file; changed names and fileds of data modules; validator fixes
