+++ tests/test_ppglm_validator.py
@@ -1,14 +1,142 @@
 from contextlib import nullcontext as does_not_raise
+from typing import Any, Optional, Union
 
+import equinox as eqx
 import jax
 import jax.numpy as jnp
 import numpy as np
 import pynapple as nap
 import pytest
 
-from nemos.pp_glm.mock_ppglm import MockPopulationPPGLM, MockPPGLM
-from nemos.pp_glm.data import X_ppglm, y_ppglm
+from nemos.base_regressor import BaseRegressor
+from nemos.glm.params import GLMParams
+from nemos.pp_glm.data import PredictorsPPGLM, SpikesPPGLM
+from nemos.pp_glm.params import PPGLMParamsWithKey
 from nemos.pp_glm.validation import PopulationPPGLMValidator, PPGLMValidator
+from nemos.regularizer import Regularizer
+from nemos.typing import UserProvidedParamsT
+
+
+class MockPPGLM(BaseRegressor):
+    """
+    Minimal PP-GLM stand-in for testing PPGLMValidator logic.
+
+    """
+
+    def __init__(
+        self,
+        n_basis_funcs: int,
+        regularizer: Optional[Union[str, Regularizer]] = None,
+        regularizer_strength: Any = None,
+        solver_name: str = None,
+        solver_kwargs: dict = None,
+    ):
+        super().__init__(
+            regularizer=regularizer,
+            regularizer_strength=regularizer_strength,
+            solver_name=solver_name,
+            solver_kwargs=solver_kwargs,
+        )
+
+        self.n_basis_funcs = n_basis_funcs
+
+        # validator is a frozen dataclass — constructed explicitly with n_basis_funcs
+        self._validator = PPGLMValidator(n_basis_funcs=self.n_basis_funcs)
+
+        self.coef_: Optional[jnp.ndarray] = None
+        self.intercept_: Optional[jnp.ndarray] = None
+
+    def _get_model_params(self) -> GLMParams:
+        return self._validator.to_model_params(
+            (
+                self.coef_,
+                self.intercept_,
+            )
+        )
+
+    def _set_model_params(self, params):
+        coef, intercept = self._validator.from_model_params(params)
+        self.coef_ = coef
+        self.intercept_ = intercept
+
+    def _check_model_is_fit(self):
+        if self.coef_ is None:
+            raise ValueError("Model is not fitted yet.")
+
+    def _model_specific_initialization(
+        self,
+        X: PredictorsPPGLM,
+        y: SpikesPPGLM,
+        **kwargs,
+    ) -> GLMParams:
+
+        empty_params = self._validator.get_empty_params(X, y)
+
+        n_neurons = jnp.unique(X.ids).size
+        n_features = int(n_neurons * self.n_basis_funcs)
+
+        init_params = eqx.tree_at(
+            lambda p: (p.coef, p.intercept),
+            empty_params,
+            (jnp.zeros(n_features), jnp.ones(n_neurons)),
+        )
+
+        return init_params
+
+    def initialize_params(
+        self,
+        X: PredictorsPPGLM,
+        y: SpikesPPGLM,
+    ) -> UserProvidedParamsT:
+        pass
+
+    def fit(self, X: PredictorsPPGLM, y: SpikesPPGLM, init_params=None):
+        """Validate inputs and set zero params; no actual optimization."""
+        fit_params = self._model_specific_initialization(X, y)
+        self._set_model_params(fit_params)
+
+    def predict(self, X):
+        self._check_model_is_fit()
+        return jnp.array(0.0)
+
+    def score(self, X, y, **kwargs):
+        self._check_model_is_fit()
+        return jnp.array(0.0)
+
+    def simulate(self, *args, **kwargs):
+        pass
+
+    def save_params(self, *args, **kwargs):
+        pass
+
+    def update(self, *args, **kwargs):
+        pass
+
+    def _initialize_optimizer_and_state(self, *args, **kwargs):
+        pass
+
+    def _log_likelihood(self, params, X, y):
+        return jnp.zeros(0.0)
+
+    def _compute_loss(
+        self,
+        params: PPGLMParamsWithKey,
+        X: PredictorsPPGLM,
+        y: SpikesPPGLM,
+        *args,
+        **kwargs,
+    ) -> jnp.ndarray:
+        return jnp.array(0.0)
+
+    def _get_optimal_solver_params_config(self, *args, **kwargs):
+        pass
+
+
+class MockPopulationPPGLM(MockPPGLM):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self._validator = PopulationPPGLMValidator(n_basis_funcs=self.n_basis_funcs)
+
 
 @pytest.fixture
 def eval_fn():
@@ -64,12 +192,12 @@ def processed_X_y(eval_fn, recording_time, valid_X_y):
     spike_times = jnp.asarray(y.t)
     y_idx = jnp.searchsorted(event_times, spike_times)
 
-    X = X_ppglm(
+    X = PredictorsPPGLM(
         times=event_times,
         ids=jnp.asarray(X.d, dtype=int),
     )
 
-    y = y_ppglm(
+    y = SpikesPPGLM(
         times=spike_times,
         ids=jnp.asarray(y.d, dtype=int),
         idx=y_idx,
@@ -160,7 +288,7 @@ class TestPPGLMValidator:
             validator.validate_consistency(params, X=X)
 
         # test that no neuron ids are skipped
-        X_skipped = X_ppglm(
+        X_skipped = PredictorsPPGLM(
             times=jnp.array([1.0, 2.0, 3.0]),
             ids=jnp.array([0, 2, 2]),  # skips id 1
         )
@@ -243,7 +371,7 @@ class TestPopulationPPGLMValidator:
 
         # non consecutive neuron ids in y
         X, _ = processed_X_y
-        y_skipped = y_ppglm(
+        y_skipped = SpikesPPGLM(
             times=jnp.array([1.0, 2.0, 3.0]),
             ids=jnp.array([0, 2, 2]),  # skips id 1
             idx=jnp.array([0, 1, 2]),
